Move uniffi AsymmEncString to bitwarden crate
